"Added proof rules for distributing basis projections into conjunctions and disjunctions. This is justified in the definition of basis projections from the FLAM paper, but did not appear in any of the rules.
Simplified the constraints on the functions in the FLAM implementation.
This was achieved by redesigning how the Label class could require constraints on the monadic context.
Previously it required that the monad class satisfied the HasCache constraint, but now we just have a completely generic constraint.

I noticed that prior to this commit, the following could would output False:

example :: FLAMIO ()
addDelegate (("a" →), ("b" →)) bot
  addDelegate (("b" ←), ("a" ←)) bot
  withStrategy [bot] $ do
    liftLIO $ modify $ (_1 . cur) .~ ("b" %)
    x <- "b" ⊑ "a"
    liftLIO $ liftIO $ print x
  return ()

runExample :: IO ()
runExample =
  evalStateT (unLIO (runFLAM example))
  (BoundedLabel { _cur = bot, _clearance = ("a" %) }, H Set.empty, [])

which was really unintuitive. The reason was that, when proving that "b" ⊑ "a",
we had to use the delegation (("a" →), ("b" →)) @ bot, but before being able to unlabel this delegation
we had to check that the label of the delegation joined with the current level flowed to the clearance.
This means we had to check that bot ⊔ "b" ⊑ "a", but we were already in the process of checking that this!
So the cycle detection mechanism kicked in and returned False in this subquery, meaning that we couldn't use the delegation!

To fix this, I introduced step indexing into the algorithm for deciding ≽. The Löb rule can then be used
to prove this "recursive" problem. Intuitively, the Löb rule says that when we need to prove "b" ≽ "a",
we are allowed to assume that this is true for subqueries. That is, we prove (⊳(a ≽ b) -> a ≽ b) -> a ≽ b, where
⊳(a ≽ b) is the assumption that a ≽ b holds for subqueries (it's called the later modality in the litterature on step-indexing).

With this rule in place the code above returns True."